localStorage service: rename filename
